chore: consolidating test logic
