+++ tests/FeaturesTests/Configuration.cs
@@ -9,22 +9,9 @@ namespace Features
     [FeatureFile(nameof(Configuration) + Constants.FeatureFileExtension)]
     public sealed class Configuration : FeatureBase
     {
-        private int? _serverIndex;
 
         public Configuration(ITestOutputHelper testOutputHelper) : base(testOutputHelper)
         {
-        }
-
-        [When(@"calling the method (\w+) without params")]
-        public async Task CallWithoutParameters(string methodName)
-        {
-            await CallMethod(methodName, null, null, _serverIndex);
-        }
-
-        [When(@"selecting the server at index (\d)")]
-        public void SelectedServerIndexIsOne(string serverIndex)
-        {
-            _serverIndex = new Nullable<int>(int.Parse(serverIndex));
         }
     }
 }
