+++ tests/FeaturesTests/HeaderParams.cs
@@ -11,19 +11,6 @@ namespace Features
         {
         }
 
-        [When(@"calling the method (\w+) with (object|array|parameters) ""(.+)""")]
-        public async Task CallMethodWithStringParameters(string methodName, string paramType, string parametersString)
-        {
-            var parameters = paramType switch
-            {
-                "object" => new object[] { _testHelper.JsonToTypeInstance("InlineObject1", parametersString) },
-                "array" => new object[] { parametersString.Split(",") },
-                _ => parametersString.Split(",")
-            };
-
-            await CallMethod(methodName, parameters);
-        }
-
         [Then(@"the request should have a header property with value (.+)")]
         public void CheckHeaderPropertyAsync(string headerProperty)
         {
