+++ tests/FeaturesTests/Objects.cs
@@ -14,34 +14,6 @@ namespace Features
         {
         }
 
-        [When(@"calling the method (\w+) and the server responds with")]
-        public async Task CallWithResponse(string methodName, DocString response)
-        {
-            await CallMethod(methodName, null, response.Content);
-        }
-
-        [Then(@"the response should be of type (\w+)")]
-        public void CheckResponseType(string type)
-        {
-            Assert.EndsWith(type, _actual.Data.GetType().Name);
-        }
-
-        [And(@"the response should have a property (id|value) with value (.+)")]
-        public void CheckResponseIdProperty(string propName, string propValue)
-        {
-            var propInfo = _actual.Data.GetType().GetProperty(propName);
-            Assert.NotNull(propInfo);
-            Assert.Equal(propValue, propInfo.GetValue(_actual.Data).ToString());
-        }
-
-        [When(@"calling the method (\w+) with (?:object|array|parameters) (.+)")]
-        public async Task CallMethodWithParameters(string methodName, string jsonTextObject)
-        {
-            var inlineObj = _testHelper.JsonToTypeInstance("InlineObject1", jsonTextObject);
-
-            await CallMethod(methodName, new object[] { inlineObj });
-        }
-
         [And("the request should have a body with value (.+)")]
         public async Task CheckRequestBody(string propValue)
         {
