+++ tests/FeaturesTests/PathParams.cs
@@ -13,25 +13,5 @@ namespace Features
         public PathParams(ITestOutputHelper testOutputHelper) : base(testOutputHelper)
         {
         }
-
-        [When(@"calling the method (\w+) with (parameters|array) ""(.+)""")]
-        public async Task CallMethodWithStringParameters(string methodName, string paramType, string parametersString)
-        {
-            var parameters = paramType switch
-            {
-                "array" => new object[] { parametersString.Split(",") },
-                _ => parametersString.Split(",")
-            };
-
-            await CallMethod(methodName, parameters);
-        }
-
-        [When(@"calling the method (\w+) with object (.+)")]
-        public async Task CallMethodWithStringObject(string methodName, string parametersString)
-        {
-            var parameters = new object[] { _testHelper.JsonToTypeInstance("InlineObject1", parametersString) };
-
-            await CallMethod(methodName, parameters);
-        }
     }
 }
